Add Makie as a weak dependency and extension module
